backend/requirements.txt
ask
flask-cors==3.0.10
psycopg2-binary==2.9.3
gunicorn==20.1.0
requests==2.26.0